Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure we should do this.   

Here's an example of why.  

Let's say you're running a live encode with a 20 frame lag and you want a new keyframe for the next frame the encoder gives you.  As of right now you can get this to happen by passing the request in as a flag.  If you change this it will mean that the keyframe will not happen for another 20 frames.  

This represents a pretty big change in behavior and while it actually makes more sense to me we are taking something away from the user.   

Ideally we should be able to change flags for the frame we are passing in and the frame we are about to encode separately.  


If we are going to make this change shouldn't we do it for vp8 as well?

Patch-set: 1
Reviewer: Gerrit User 1000963 <1000963@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1