Allow OVERLAY frame using show_exsiting_frame
ARF with zero strength temporal filter can be reused by setting the
show_existing_frame flag = 1, and in this case, there is no need to
refresh the reference frame buffer. However, we used the flag
"refresh_golden_frame" as the identifier for the starting point of a gf
group.

Two new flags "is_arf_filter_off" and "is_gf_group_begin" to track if
the zero strength filter is used, and decouple the dependencies of some
gf related function from "refresh_golden_frame."

Two extra flags are added in the configure file:
show_existing_arf: enable encoder to use the show_existing_frame for ARF
flag if the temporal filter strength is zero.
str_zero: force the temporal filter strength for ARF to be zero.

Change-Id: I25971a760f6e1638d5147fe30488c48125512b1a