Allow OVERLAY frames to use the show_exsiting_frame flag
ARF with zero strength temporal filter can be reused by setting the
show_existing_frame = 1, and in this case, there is no need to
refresh the reference frame buffer. However, we used the flag
"refresh_golden_frame" as the identifier for the starting point of a gf
group.

Two new flags "is_arf_filter_off" and "is_gf_group_start" are used to track
if the zero strength filter is used for a overlay frame. Therefore, we
can still update rate-control parameter properly.

Change-Id: I25971a760f6e1638d5147fe30488c48125512b1a