Upload segmentation map to the cpi->alt_ref_aq
    I   use  here   simple   though   reasonable  approach   creating
segmentation map  as a number  of nonzero frames blurred  together to
produce current  pixel.  I also experimented  with non-binary weights
(actually 0, 1 and 2 like in the superframe creation process itself),
but results are similar in average.
    Right now,  I am trying  to consider larger delta  quantizers for
the low  error since  low errors mean  that these  blocks compensates
well from previous frame or key frame  and we may save bitrate in the
first, but not last case. Unfortunately,  it just completely does not
work at the moment.

Change-Id: Id490079daf086cf3b7b54eab3f0a92461b8dce45