Update patch set 2
Patch Set 2: -Code-Review

The cause of the failure might be this patch https://chromium-review.googlesource.com/#/c/358492/25/vp10/common/scan.c, in which vp10_default_scan_orders[] was deleted because it is not used in the codec. But the unit test uses it. Probably adding it back is a quick solution.

Patch-set: 2
Label: Code-Review=0