Implement multi-ARFs' to emulate heirachical B frames
Insert multiple arfs in a gf group to emulate heirachical B
frames structure. A new config flag "ext-arfs" is added to
enable/disable multiple ARFs'.

When the internal ARFs' is forced to be non-filtered and use
the show_existing_frame flag, an unusual speed-up is observed.
One reason is that the Q value will be changed, and encoder
will not trigger the recode process.

ToDo:
Determine the rate controller for newly inserted alt_refs'.

Change-Id: I62c32e1b0f25b978484dd113b319bebcd959bf60