Implement multi-arfs to emulate heirachical B frames
Insert multiple arfs in a gf group to emulate heirachical B
frames structure.

The basical frameworks is done, but there are some works need to
be done, and there is an unexpected speed-up when showing exsiting
extra arfs' as overlay frames, which need to be investigated.

ToDo:
Determine the rate controller for newly inserted alt_refs'.

Change-Id: I62c32e1b0f25b978484dd113b319bebcd959bf60