Align thread entry point stack
_beginthreadex does not align the stack on 16-byte boundary as expected by gcc.

On x86 targets, the force_align_arg_pointer attribute may be applied to
individual function definitions, generating an alternate prologue and epilogue
that realigns the run-time stack if necessary. This supports mixing legacy
codes that run with a 4-byte aligned stack with modern codes that keep a 16-byte
stack for SSE compatibility.
https://gcc.gnu.org/onlinedocs/gcc/x86-Function-Attributes.html

Change-Id: Ie4e4ab32948c238fa87054d5664189972ca6708e
Signed-off-by: Aleksey Vasenev <margtu-fivt@ya.ru>