Use hex representation for sinpi8sqrt2
The value 35468 changes sign when stored in int16_t:
implicit conversion from 'int' to 'int16_t' (aka 'short')
changes value from 35468 to -30068

The value is used in a multiplication which treats it as unsigned
despite the type being signed. I suspect the value is cast to s32
without sign extension for some reason.

BUG=b/28027557

Change-Id: I0c5ce09bea170fe08061856c2af6f841a557e0c3