Change the B-frame coding structure.
Originally we can have a backward reference frame right before an
overlay frame, which might be unnecessary since we already has a
quality backward reference frame (ARF).
This patch forbids such a coding structure and improves the RD
performance by 0.086% in Avg in the lowres dataset.
The biggest drop is 0.028% for the motherdaughter_cif sequence, and
significant gains are obtained for the following sequences:

mobisode2_240p: 0.563%
keiba_240p: 0.440%
bus_cif: 0.336%
soccer_cif: 0.333%

Change-Id: Ic94f648ba8e52eb0014933d484fb247610a9ae05