Allow LF_UPDATE type of frames to use BWDREF
Originally, only bi-pred type of frames can use BWDREF. When
extra alt-refs are inserted in a gf group, the closest alt-ref
serves as ALTREF for the frames within the corresponding
subgroup. Therefore, the original alt-ref can be used as BWDREF
for the LF_UPDATE type of frames.

This patch further swaps the virtual indices of BWDREF and ALTREF
for those frames whose BWDREF is farther than ALTREF. As a result,
the BWDREF is always the closet backward reference frame, and the
ALTREF is the farther one.

It improves the average RD performance by

0.132% in lowres, and
0.030% in midres.

The overall gains for the ext-refs compared to the baseline are

5.486% in lowres, and
4.666% in midres.

Change-Id: I22e4e5f378f19c4c89196a0a5e9214adb46c3428