vp8 : threading remove unnecessary pause.
The pause right before a sleep is at best unnecessary.  As the pause
instruction is in itself like a sleep.

in intel's article : benefitting-power-and-performance-sleep-loops
Pause is described as providing major benefits over sleep loops.
However testing in this code base doesn't bare that out.   I tried the
following code on 4 different machines with no apparent gains,  with
values for tries from 0 to 4000,  but typically at 10 to 15 with the
pause instruction and 1000- 2000 without it :

Here's a snippet of code I tried:

  const int kTries = 10;
  int i;
  while (1) {
    if (mb_col <= read_or_zero(mutex, last_row_mb_col) - nsync) return;
    for (i = 0; i < kTries; ++i) {
      x86_pause_hint();
      if (mb_col <= read_or_zero(mutex, last_row_mb_col) - nsync) return;
    }
    thread_sleep(0);
  }

read_or_zero does a try lock or returns a 0 if it can't get it.

Change-Id: Ia5e0ac57834bef1d2cf3ef287167eaa96087c43d