Restore vp8_sixtap_predict4x4_neon
This function was removed when clang started introducing alignment hints
which caused the 32 bit vld1_lane_u32/vst1_lane_u32 to fail:
https://llvm.org/bugs/show_bug.cgi?id=24421

The load has been rendered safe with an implementation ~indiscernible
performance-wise that uses _u8 and over-reads just a touch.

The store, when unaligned, has a version that is ~25% slower but safe.
Since the crash reports were somewhat few and far between, this appears
to be a fairly rare condition. As a result, it checks for alignment and
uses the previous cast when possible.

Overall it is ~10x faster than C and ~5x faster than the armv6 version
which was removed recently.

BUG=webm:817
BUG=webm:1273

Change-Id: I1e490e94453e0872151fe0dafb05557463f6247d