Restore vp8_bilinear_predict4x4_neon
This function was removed when clang started introducing alignment hints
which caused the 32 bit vld1_lane_u32/vst1_lane_u32 to fail:
https://llvm.org/bugs/show_bug.cgi?id=24421

The load has been rendered safe with an implementation ~indiscernible
performance-wise that uses _u8 and over-reads just a touch.

Like the sixtap, there is a substantial penalty for the unaligned
stores. However, the function is much smaller so the effect is more
pronounced.

Unlike the sixtap, there was no existing test. Copied the sixtap.
Unfortunately the ssse3 version is failing the test. Will investigate.

Overall it is ~5x faster than C.

BUG=webm:892
BUG=webm:1273
BUG=webm:1287

Change-Id: Icf7167189391b46202f47233bb585c24c42bcc36