Enable tile-adaptive restoration
Includes a major refactoring to support tile-adaptive switchable
restoration. The framework can be readily extended to add more
restoration schemes in the future. Also includes various cleanups
and fixes.

The framework now allows restoration to be conducted on tiles
such that each tile can be either left unrestored, or use bilateral
or use wiener.
There is not much change in coding efficiency yet, but further
enhancements will be added to improve coding efficiency.

Change-Id: I5ebedb04785ce1ef6f324abe209e925c2d6cbe8a