Split the vp9_highbitdepth option into encode and decode parts
DO NOT MEGRE THIS PATCH

This is just an experiment to measure the impact of splitting
highbitdepth into two parts.  It saves about 486kBytes of
code size when we want to encode with only 8bits, but decode with 8/10/12.

However, I haven't run any tests on it so it is not likely
to be correct.

Change-Id: I5853c27dc58afcd72c6d5265de7fa2ccdb708eaf