Hook up vp8_diamond_search_sad_sse3
The original commit never set any 'specialize' line:
61311e61039c300ae872ccba22304e9e60dc0205

It appears to use the sadx4 function to speed up the sad diamond search.
There are no sse3 versions of the sx4df functions, but there are sse2
versions.

There are neon and msa versions of the vpx_sad16x16x4d but not any of
the smaller versions. Perhaps if they existed this function could be
expanded to use them.

Change-Id: I936d7d6b1a3ff6dcd5a4d2322272708c47cdec13