Update patch set 7
Patch Set 7:

(1 comment)

Also, I'm having a hard time determining the semantics when multiple traps are installed:
- tools_common.sh traps EXIT and executes cleanup()
- stress.sh (the one you haven't uploaded yet; yes, I'm reviewing unpublished code... evil, no?) traps INT and TERM

I think you're going to have to call cleanup() manually from your trap function to be certain it executes if/when stress.sh is run with the other tests to avoid leaving a bunch of temp files in $TEMP.

In a normal run without an error dropping us into the trap function I think it's going to be fine.

Alternatively, you could just download the input to the temp dir and let it get deleted as part of the example test normal clean up step. That just sucks because you'll need to download the file every time you run stress.sh.

Anyway, I guess this sort of forces the question: Should stress.sh download a file and leave it no the disk, or not?

Patch-set: 7
Label: Code-Review=0