Update patch set 2
Patch Set 2:

I have integrated some parts of this into the other patch:
https://chromium-review.googlesource.com/#/c/408558/
My preference is to first get to a state where the gmtype is not stored at all and the tranformation type is always derivable from the matrix.
Later we can store the gmtype or wmtype just as a matter of convenience or efficiency.

Patch-set: 2
Label: Code-Review=0