Fix clpf + ext_partition and dering + ext_partition
In both cases, the data to be signalled (for clpf/dering) is
stored at the topmost superblock level, but previously this data
would not be signalled for 128x128 superblocks.

For dering, it was also assumed that all superblocks were of the
maximum possible size. When ext_partition was enabled and we
chose to use a 64x64 superblock, this resulted in an out-of-bounds
access to cm->mi_grid_visible, leading to a segfault.

Change-Id: Ie0a56201717069a7060f9331dc5817a32d746de7