Update patch set 4
Patch Set 4:

Do I need to do anything more here, or is this good for review?

Note that in the original code, if shortcut is not set, there is a continue to the next iteration. Thus, the latter test on shortcut is pointless, since shortcut is always set at that point.

Patch-set: 4
Label: Code-Review=0