Update patch set 1
Patch Set 1:

cpi->sf.recode_loop is a speed feature. For cpu-used = 0 ~ 2, the current frame's size is actually calculated here for better decision making. However, the encoder could be somewhat slow because of the calling of vp9_pack_bitstream. For higher cpu-used setting(namely, faster encoder with a little worse quality), the previous frame's size is used instead. So, this is a designed feature, not a bug.

Patch-set: 1