Update patch set 21
Patch Set 21:

The CL looks good, but for some reason I can't +2 it.

After this I think we should make a follow on CL that adds a GET_LEVEL call to return the current level. This way applications can check the max bitrate after the encoding is done.

Patch-set: 21
Label: Code-Review=0