Update patch set 2
Patch Set 2: Code-Review+1

As a further note the "abrupt" relaxation of the active maxq does not force the codec to use that maximum Q value, it simply increases the range of values available in the recode loop.  The main issue at the moment I suspect is that once the extended range is available the recode loop will try and use that Q range to get closer to the nominal rate target for the frame rather than to just sneak under the upper limit value. This could mean that a frame of lower complexity that does not exceed the maximum rate will end up getting more bits than one that triggers the relaxation and then uses a much higher Q to get closer to its nominal rate target. Ideally in this situation therefore we would want to relax Q only just far enough to get under the critical rate limit.

Patch-set: 2
Label: Code-Review=+1