Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> update the README and bump the version in iosbuild.sh, I think the idea there was to track the latest. really we should revisit this and just target the user's selected sdk, allowing them to manipulate cflags as necessary

In iosbuild.sh, if --macosx flag is used, it will choose one specific target (right now is darwin15 which comes with macOS 10.11). If this is updated to darwin16, for those who haven't upgraded macOS to the latest, the build will fail. Am I miss anything? 
Actually I've tested that even using darwin15 on macOS 10.12, there is no speed difference observed. Since I guess macOS 10.12 is not so popular today, is it better to keep it on darwin15 in iosbuild.sh?

Patch-set: 2
Label: Code-Review=0