Update patch set 1
Patch Set 1:

> why didn't anything break?

Because it produces the same offset - for the existing uses. When I wanted to use this for quantize ssse3, I ran into problems with ncoeffs.

This just rebalances the multiplications.

Before it was 1 * 32
Now it is 8 * 4.

This reflects the actual number of elements (8).

Patch-set: 1
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>