Experiment to improve first pass bit distribution
** Local debug stats output enabled.

Step 1: (Commit 1d8bdb...)
Force recode on all frames and allow full Q range. This forces the codec
to accurately track the frame to frame bit allocation. This change causes a
big drop in metrics.

Step 2:

Change to the per frame boost calculation. Add a minimum boost but greatly
reduce the dependence on the error score. Now only quite low motion clips
will get a substantial extra boost above the baseline level and the total boost
will depend more on the number of frames in the group. In most case the bit
allocation for alt ref frames is substantially reduced.

Improves psnr, psnr-hvs and ssim metrics for accurate tracking substantially.
(eg. Low res set psnr 9.7%).

However this does not translate yet to big gains in normal vbr mode. This may
be partly due to dependencies elsewhere in the code on the arf boost number
that need to be tuned.

Change-Id: Icd5c21b332acd5279a342e7f7688abcae2a302ef