VP9 motion vector unit test
To prevent the motion vector out of range bug, added a motion vector unit
test in VP9. In the 4k video encoding, always force to use extreme motion
vectors and also encourage to use INTER modes. In the decoding, check if
the motion vector is valid, and also check the encoder/decoder mismatch.

Change-Id: I0a880bd847dad8a13f7fd2012faf6868b02fa3b4