VP9 motion vector unit test
To prevent the motion vector out-of-range bug, added a motion vector unit
test in VP9. In the 4k video encoding, always force to use extreme motion
vectors and also encourage to use NEWMV mode. In the decoding, check if the
MV is valid, and also turn on the encoder/decoder mismatch checking.

Change-Id: I0a880bd847dad8a13f7fd2012faf6868b02fa3b4