Experiment. Store first pass errors as per MB values.
Most existing first pass stats are stored in a form normalized to a
macro-block scale. However the error scores for intra / inter etc were
stored as frame level values but mainly used as MB level values.

This change change fixes that. Normalized per MB values make
comparisons between different formats easier and in any case this is
usually what is wanted.

An change in results should be limited to slight differences in rounding.

Change-Id: Ic2d57980398c99ade2b7380e3e6ca6b32186901f