Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> (1 comment)
> 
> +2 with the one comment on possible variation.

I thought about this. If we want to make this change only for multi-thread case, then, to be safe, we may need to keep ex_max limitation for single-thread case. Since single-thread and multi-thread encoding share the same code. Then, it means the 2 counts still exist at multi-thread case and we just don't use them. In this way, we have to keep the synchronizations around the counts, which causes the code more complicated and may introduce more overhead because of the synchronizations. Anyway, for hd clips, we may want to use multi-threaded encoding most of time, maybe it is better to keep the 2 case consistent.

Patch-set: 2
Reviewer: Gerrit User 1000884 <1000884@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>