Update patch set 6
Patch Set 6:

I believe the original behaviour was correct when using arf groups. The frame 1 in your example is a key frame and a special case. Frame 5 is the overlay frame for arf 1 so the actual arf group of 4 frames is frames 2,3,4,5 where the arf1 coincides temporally with frame 5. In the modified example that you have now reverted a min group length of 4 would actually result in 3 real frames plus an arf per group not an arf group containing 4 real frames. Where there are no ARFs and we are just using GF references the case is different because the GF at the start of each group is itself a real frame.

Patch-set: 6
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>